remove clients, todos
