--- duva/src/clients/mod.rs
@@ -1,2 +1,0 @@
-// TODO move this module to presentation
-pub mod authentications;
