+++ duva/src/domains/peers/connected_peer_info.rs
@@ -7,7 +7,6 @@ use super::{
 
 #[derive(Debug, Clone, Default)]
 pub(crate) struct ConnectedPeerInfo {
-    // TODO repl_id here is the leader_replid from connected server.
     pub(crate) id: PeerIdentifier,
     pub(crate) replid: ReplicationId,
     pub(crate) hwm: u64,
