+++ duva/src/presentation/clients/mod.rs
@@ -3,5 +3,7 @@ pub mod controller;
 pub mod parser;
 pub mod request;
 pub mod stream;
+pub use authenticate::AuthRequest;
+pub use authenticate::AuthResponse;
 pub(crate) use authenticate::authenticate;
 pub(crate) use controller::ClientController;
