+++ duva/src/presentation/clients/stream.rs
@@ -45,8 +45,6 @@ impl ClientStreamReader {
         sender: Sender<QueryIO>,
     ) {
         loop {
-            //TODO check on current mode of the node for every query? or get notified when change is made?
-
             match self.extract_query().await {
                 Ok(requests) => {
                     let results = match handler.maybe_consensus_then_execute(requests).await {
