+++ duva/tests/cluster_ops/test_lazy_discovery.rs
@@ -25,7 +25,6 @@ async fn test_lazy_discovery_of_leader() -> anyhow::Result<()> {
     let cluster_info = other_h.send_and_get("CLUSTER INFO".as_bytes(), 1).await;
     assert_eq!(cluster_info.first().unwrap(), "cluster_known_nodes:0");
 
-    // TODO can be flaky?
     assert_eq!(
         other_h.send_and_get("KEYS *".as_bytes(), 2).await,
         vec!["0) \"other2\"", "1) \"other\""]
